# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Urological Review System API
  description: Backend API for managing urological residency application reviews
  version: 1.0.0
  contact:
    name: Urological Review System
    email: support@example.com
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.yourproductiondomain.com
    description: Production server
paths:
...
  /api/applicants:
    get:
      summary: Get all applicants
      description: Retrieve all applicants in the system
      responses:
        '200':
          description: Successfully retrieved applicants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DatabaseApplicant'
    post:
      summary: Create a new applicant
      description: Add a new applicant to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicantRequest'
      responses:
        '201':
          description: Applicant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DatabaseApplicant'
        '400':
          description: Validation error
        '409':
          description: Applicant already exists
...
  /api/reviews:
    get:
      summary: Get reviews
      description: Retrieve reviews with optional filtering
      parameters:
        - name: applicant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: reviewer_name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DatabaseReview'
    post:
      summary: Create a review
      description: Submit a new review for an applicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created successfully
        '400':
          description: Validation error
        '409':
          description: Review already exists
  /api/reviews/applicant/{applicantId}:
    get:
      summary: Get reviews for applicant
      description: Retrieve all reviews for a specific applicant
      parameters:
        - name: applicantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DatabaseReview'
  /api/reviews/final-selections:
    get:
      summary: Get all final selections
      description: Retrieve all final administrative decisions
      responses:
        '200':
          description: Successfully retrieved final selections
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DatabaseFinalSelection'
    post:
      summary: Create/update final selection
      description: Make or update a final administrative decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFinalSelectionRequest'
      responses:
        '201':
          description: Final selection created/updated successfully
...
components:
  schemas:
    ApplicantCategory:
      type: string
      enum: [regular, i-sub]
    AdminDecision:
      type: string
      enum: [Selected, "Not Selected", Pending]
    DecisionType:
      type: string
      enum: ["Definitely Interview", Maybe, "Do Not Interview"]
    DatabaseApplicant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/ApplicantCategory'
        details:
          type: string
        site_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DatabaseReview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicant_id:
          type: string
          format: uuid
        reviewer_name:
          type: string
        preference:
          type: integer
          minimum: 1
          maximum: 5
        pressure:
          type: integer
          minimum: 1
          maximum: 5
        underserved:
          type: integer
          minimum: 1
          maximum: 5
        leadership:
          type: integer
          minimum: 1
          maximum: 5
        academic:
          type: integer
          minimum: 1
          maximum: 5
        research:
          type: integer
          minimum: 1
          maximum: 5
        personal:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: string
        decision:
          $ref: '#/components/schemas/DecisionType'
        total_score:
          type: integer
        site_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DatabaseFinalSelection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicant_id:
          type: string
          format: uuid
        admin_decision:
          $ref: '#/components/schemas/AdminDecision'
        selection_reason:
          type: string
        average_score:
          type: number
        reviewer_count:
          type: integer
        site_name:
          type: string
        decided_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateApplicantRequest:
      type: object
      required:
        - external_id
        - name
      properties:
        external_id:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/ApplicantCategory'
        details:
          type: string
...
    CreateReviewRequest:
      type: object
      required:
        - applicant_id
        - reviewer_name
      properties:
        applicant_id:
          type: string
          format: uuid
        reviewer_name:
          type: string
        preference:
          type: integer
          minimum: 1
          maximum: 5
        pressure:
          type: integer
          minimum: 1
          maximum: 5
        underserved:
          type: integer
          minimum: 1
          maximum: 5
        leadership:
          type: integer
          minimum: 1
          maximum: 5
        academic:
          type: integer
          minimum: 1
          maximum: 5
        research:
          type: integer
          minimum: 1
          maximum: 5
        personal:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: string
        decision:
          $ref: '#/components/schemas/DecisionType'
    CreateFinalSelectionRequest:
      type: object
      required:
        - applicant_id
        - admin_decision
      properties:
        applicant_id:
          type: string
          format: uuid
        admin_decision:
          $ref: '#/components/schemas/AdminDecision'
        selection_reason:
          type: string
...